+++ apps/web/src/components/common/attach-images/attach-images-thumbnail.tsx
