+++ apps/web/src/components/common/auction-card/badge-bid-status.tsx
