+++ apps/web/src/components/common/auction-card/card.tsx
@@ -2,7 +2,7 @@
 
 import Link from 'next/link';
 
-import BadgeBidStatus from '@/components/common/auction-item/badge-bid-status';
+import BadgeBidStatus from '@/components/common/auction-card/badge-bid-status';
 import Thumbnail from '@/components/ui/thumbnail';
 import { Badge } from '@/components/ui/badge';
