+++ apps/web/src/views/CreateAcutionPage.tsx
@@ -2,7 +2,7 @@
 
 import { CircleAlert } from 'lucide-react';
 import CreateAuctionForm from '@/components/auction-create/form-create';
-import Notice from '@/components/common/auction-form/notice';
+import Notice from '@/components/common/messages/notice';
 import { Button } from '@/components/ui';
 import useCreateAuction from '@/hooks/auction/useCreateAuction';
